Adding ACClient craziness
